requirements.txt
pytesseract
pillow
pyautogui
keyboard